app/schemas/auth.py
from pydantic import BaseModel

class Token(BaseModel):
    access_token: str
    token_type: str

class TokenData(BaseModel):
    email: str = None

class UserLogin(BaseModel):
    email: str
    password: str

class RefreshToken(BaseModel):
    refresh_token: str